consistent changes mirror /import version

+++ server/battleUtils.js
@@ -1,19 +1,33 @@
 import _ from 'underscore';
+
 import { BUFFS } from '../imports/constants/buffs/index';
 
 // why is this a duplicate of /imports/battleUtils.js ?
 
 export const removeBuff = function removeBuff({ target, buff, caster, actualBattle }) {
-  const buffConstants = BUFFS[buff.id];
-  // Quick sort of buffs to ascending by duration
-  // target.buffs = _.sortBy(target.buffs, 'duration');
-  if (buffConstants.events.onRemove) {
-    buffConstants.events.onRemove({ buff, target, caster, actualBattle });
+  if (target && buff && buff.id) {
+    if (target.removeBuff) {
+      actualBattle = actualBattle || target.battleRef;
+      
+      const buffConstants = BUFFS[buff.id];
+      // Quick sort of buffs to ascending by duration
+      // target.buffs = _.sortBy(target.buffs, 'duration');
+      if (buffConstants.events.onRemove) {
+        buffConstants.events.onRemove({ buff, target, caster, actualBattle });
+      }
+      target.removeBuff(buff);
+    }
   }
-  target.removeBuff(buff);
-}
+};
 
 export const addBuff = function addBuff({ buff, target, caster, actualBattle }) {
+  if (!target || !target.addBuff || !buff || !buff.data) {
+    return;
+  }
+  //console.log(`applying ${BUFFS[buff.id].name} from ${caster.name} to ${target.name}`);
+
+  actualBattle = actualBattle || target.battleRef;
+  
   let existingBuff;
   if (!buff.data.allowDuplicates) {
     // Make sure there is no existing buff like this
@@ -54,4 +68,41 @@ export const addBuff = function addBuff({ buff, target, caster, actualBattle }) 
     newBuff.onApply({ buff: newBuff, target, caster, actualBattle });
     newBuff.data.didApply = true;
   }
-}
+  
+  return newBuff;
+};
+
+export const removeBuffById = function removeBuffById({target, caster, buffId, actualBattle}) {
+  try {
+    actualBattle = actualBattle || target.battleRef;
+    
+    let buffToRemove = target.findBuff(buffId);
+    if (buffToRemove) {
+      removeBuff({ buff: buffToRemove, target, caster, actualBattle });
+      return true;
+    }
+  } catch (err) {
+  }
+  return false;
+};
+
+export const removeBuffWithMessage = function removeBuffWithMessage({target, caster, buff, buffId, actualBattle, messageOptions}) {
+  try {
+    actualBattle = actualBattle || target.battleRef;
+    
+    if (buff) {
+      removeBuff({target, buff, caster, actualBattle});
+      if (messageOptions) {
+        target.tickMessage(messageOptions.label, messageOptions.color);
+      }
+      return true;
+    } else if (removeBuffById({target, caster, buffId, actualBattle})) {
+      if (messageOptions) {
+        target.tickMessage(messageOptions.label, messageOptions.color);
+      }
+      return true;
+    }
+  } catch (err) {
+  }
+  return false;
+};
